+++ packages/eslint-config/index.js
@@ -9,9 +9,9 @@ module.exports = {
     es6: true,
     node: true,
   },
-  extends: ['plugin:import/recommended', 'airbnb', 'prettier'],
   parser: '@babel/eslint-parser',
-  plugins: ['eslint-plugin-react-hooks'],
+  extends: ['airbnb', 'prettier'],
+  plugins: ['react-hooks'],
   settings: {
     react: {
       version: 'detect',
@@ -67,6 +67,14 @@ module.exports = {
     'import/no-extraneous-dependencies': 'off',
     // The code is already coupled to webpack. Prefer explicit coupling.
     'import/no-webpack-loader-syntax': 'off',
+    'import/order': [
+      'error',
+      {
+        groups: ['builtin', 'external', 'internal', 'parent', 'sibling', 'index'],
+        pathGroups: [{ pattern: '~/**', group: 'internal' }],
+        'newlines-between': 'never',
+      },
+    ],
 
     'jsx-a11y/label-has-associated-control': 'off',
     'jsx-a11y/no-autofocus': 'off',