+++ packages/oui/test/TestProvider.js
@@ -1,6 +1,5 @@
 import * as React from 'react'
 import PropTypes from 'prop-types'
-import { ThemeProvider as EmotionThemeProvider } from 'emotion-theming'
 import { createTheme, ThemeProvider } from '@mui/material/styles'
 
 const theme = createTheme({})
@@ -9,15 +8,9 @@ function TestProvider(props) {
   const { children, ...other } = props
 
   return (
-    <EmotionThemeProvider
-      // Bug: Custom theme not propagated within Storybook.js
-      // https://github.com/mui-org/material-ui/issues/24282#issuecomment-859393395
-      theme={theme}
-    >
-      <ThemeProvider theme={theme} {...other}>
-        {children}
-      </ThemeProvider>
-    </EmotionThemeProvider>
+    <ThemeProvider theme={theme} {...other}>
+      {children}
+    </ThemeProvider>
   )
 }