+++ packages/react-centra-checkout/src/index.ts
@@ -2,8 +2,11 @@
 export { default as ApiClient } from './ApiClient'
 export * from './ApiClient'
 
+export { default as CentraSelectionContext } from './Context'
+export * from './Context'
+
 export { default as PaymentEmbed } from './PaymentEmbed'
 export * from './PaymentEmbed'
 
-export { default as CentraSelectionContext } from './Context'
-export * from './Context'
+export { default as ShipwalletEmbed } from './ShipwalletEmbed'
+export * from './ShipwalletEmbed'