+++ packages/react-centra-checkout/src/internal/HtmlEmbed/index.tsx
@@ -1,16 +1,19 @@
 import * as React from 'react'
 
-interface PaymentEmbedHtmlProps extends React.ComponentProps<'div'> {
+interface HtmlEmbedProps extends React.ComponentProps<'div'> {
   html: string
 }
 
-function PaymentEmbedHtml(props: PaymentEmbedHtmlProps): React.ReactElement {
+function HtmlEmbed(props: HtmlEmbedProps): React.ReactElement {
   const { html, ...other } = props
 
-  const [node, setNode] = React.useState<HTMLDivElement | null>(null)
+  const ref = React.useRef<HTMLDivElement>(null)
 
   // We need to reload scripts in the html in order for them to run
   React.useEffect(() => {
+    const createdScripts: HTMLScriptElement[] = []
+    const node = ref.current
+
     if (node && html) {
       const scripts = node.getElementsByTagName('script')
 
@@ -24,13 +27,19 @@ function PaymentEmbedHtml(props: PaymentEmbedHtmlProps): React.ReactElement {
           scriptEl.appendChild(textNode)
         }
 
+        script.remove()
+        createdScripts.push(scriptEl)
         node.appendChild(scriptEl)
       })
     }
-  }, [node, html])
+
+    return () => {
+      createdScripts.forEach((script) => script.remove())
+    }
+  }, [html])
 
   // eslint-disable-next-line react/no-danger
-  return <div ref={setNode} dangerouslySetInnerHTML={{ __html: html }} {...other} />
+  return <div ref={ref} dangerouslySetInnerHTML={{ __html: html }} {...other} />
 }
 
-export default PaymentEmbedHtml
+export default HtmlEmbed