+++ packages/eslint-config-typescript/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/noaignite/accelerator/compare/@noaignite/eslint-config-typescript@0.2.2...@noaignite/eslint-config-typescript@0.2.4) (2022-06-01)
+
+**Note:** Version bump only for package @noaignite/eslint-config-typescript
+
+
+
+
+
+## [0.2.3](https://github.com/noaignite/accelerator/compare/@noaignite/eslint-config-typescript@0.2.2...@noaignite/eslint-config-typescript@0.2.3) (2022-04-12)
+
+**Note:** Version bump only for package @noaignite/eslint-config-typescript
+
+
+
+
+
+## [0.2.2](https://github.com/noaignite/accelerator/compare/@noaignite/eslint-config-typescript@0.2.1...@noaignite/eslint-config-typescript@0.2.2) (2022-04-12)
+
+**Note:** Version bump only for package @noaignite/eslint-config-typescript
+
+
+
+
+
+## [0.2.1](https://github.com/noaignite/accelerator/compare/@noaignite/eslint-config-typescript@0.2.0...@noaignite/eslint-config-typescript@0.2.1) (2022-04-06)
+
+
+### Bug Fixes
+
+* **import:** remove plugin:import/recommended extend ([097b2ce](https://github.com/noaignite/accelerator/commit/097b2cebb504c2c05a3b678d54b4ecea9938273a))
+
+
+
+
+
+# [0.2.0](https://github.com/noaignite/accelerator/compare/@noaignite/eslint-config-typescript@0.1.1...@noaignite/eslint-config-typescript@0.2.0) (2022-03-22)
+
+
+### Features
+
+* don't use overrides ([f073ead](https://github.com/noaignite/accelerator/commit/f073ead3523d9213b5e54d547b36a012ebdd390c))
+
+
+
+
+
 ## [0.1.1](https://github.com/noaignite/accelerator/compare/@noaignite/eslint-config-typescript@0.1.0...@noaignite/eslint-config-typescript@0.1.1) (2022-03-10)
 
 **Note:** Version bump only for package @noaignite/eslint-config-typescript